// src/types/MiscellaneousItemData.ts
export interface MiscellaneousItemData {
    name: string;
    ['*name']: string;
    szFlavorText: string;
    compactsave: string;
    version: string;
    level: string;
    levelreq: string;
    rarity: string;
    spawnable: string;
    speed: string;
    nodurability: string;
    cost: string;
    ['gamble cost']: string;
    code: string;
    alternategfx: string;
    namestr: string;
    ['auto prefix']: string;
    component: string;
    invwidth: string;
    invheight: string;
    hasinv: string;
    gemsockets: string;
    gemapplytype: string;
    flippyfile: string;
    invfile: string;
    uniqueinvfile: string;
    special: string;
    Transmogrify: string;
    TMogType: string;
    TMogMin: string;
    TMogMax: string;
    useable: string;
    throwable: string;
    type: string;
    type2: string;
    dropsound: string;
    dropsfxframe: string;
    usesound: string;
    unique: string;
    transparent: string;
    transtbl: string;
    lightradius: string;
    belt: string;
    autobelt: string;
    stackable: string;
    minstack: string;
    maxstack: string;
    spawnstack: string;
    quest: string;
    questdiffcheck: string;
    missiletype: string;
    spellicon: string;
    pSpell: string;
    state: string;
    cstate1: string;
    cstate2: string;
    len: string;
    stat1: string;
    calc1: string;
    stat2: string;
    calc2: string;
    stat3: string;
    calc3: string;
    spelldesc: string;
    spelldescstr: string;
    spelldesccalc: string;
    durwarning: string;
    qntwarning: string;
    gemoffset: string;
    BetterGem: string;
    bitfield1: string;
    CharsiMin: string;
    CharsiMax: string;
    CharsiMagicMin: string;
    CharsiMagicMax: string;
    CharsiMagicLvl: string;
    GheedMin: string;
    GheedMax: string;
    GheedMagicMin: string;
    GheedMagicMax: string;
    GheedMagicLvl: string;
    AkaraMin: string;
    AkaraMax: string;
    AkaraMagicMin: string;
    AkaraMagicMax: string;
    AkaraMagicLvl: string;
    FaraMin: string;
    FaraMax: string;
    FaraMagicMin: string;
    FaraMagicMax: string;
    FaraMagicLvl: string;
    LysanderMin: string;
    LysanderMax: string;
    LysanderMagicMin: string;
    LysanderMagicMax: string;
    LysanderMagicLvl: string;
    DrognanMin: string;
    DrognanMax: string;
    DrognanMagicMin: string;
    DrognanMagicMax: string;
    DrognanMagicLvl: string;
    HraltiMin: string;
    HraltiMax: string;
    HraltiMagicMin: string;
    HraltiMagicMax: string;
    HraltiMagicLvl: string;
    AlkorMin: string;
    AlkorMax: string;
    AlkorMagicMin: string;
    AlkorMagicMax: string;
    AlkorMagicLvl: string;
    OrmusMin: string;
    OrmusMax: string;
    OrmusMagicMin: string;
    OrmusMagicMax: string;
    OrmusMagicLvl: string;
    ElzixMin: string;
    ElzixMax: string;
    ElzixMagicMin: string;
    ElzixMagicMax: string;
    ElzixMagicLvl: string;
    AshearaMin: string;
    AshearaMax: string;
    AshearaMagicMin: string;
    AshearaMagicMax: string;
    AshearaMagicLvl: string;
    CainMin: string;
    CainMax: string;
    CainMagicMin: string;
    CainMagicMax: string;
    CainMagicLvl: string;
    HalbuMin: string;
    HalbuMax: string;
    HalbuMagicMin: string;
    HalbuMagicMax: string;
    HalbuMagicLvl: string;
    MalahMin: string;
    MalahMax: string;
    MalahMagicMin: string;
    MalahMagicMax: string;
    MalahMagicLvl: string;
    LarzukMin: string;
    LarzukMax: string;
    LarzukMagicMin: string;
    LarzukMagicMax: string;
    LarzukMagicLvl: string;
    DrehyaMin: string;
    DrehyaMax: string;
    DrehyaMagicMin: string;
    DrehyaMagicMax: string;
    DrehyaMagicLvl: string;
    JamellaMin: string;
    JamellaMax: string;
    JamellaMagicMin: string;
    JamellaMagicMax: string;
    JamellaMagicLvl: string;
    ['Source Art']: string;
    ['Game Art']: string;
    Transform: string;
    InvTrans: string;
    SkipName: string;
    NightmareUpgrade: string;
    HellUpgrade: string;
    mindam: string;
    maxdam: string;
    PermStoreItem: string;
    multibuy: string;
    Nameable: string;
    ['*eol']: 0
}