// src/types/ArmorItemData.ts
export interface ArmorItemData {
    name: string;
    version: string;
    compactsave: string;
    rarity: string;
    spawnable: string;
    minac: string;
    maxac: string;
    absorbs: string;
    speed: string;
    reqstr: string;
    block: string;
    durability: string;
    nodurability: string;
    level: string;
    levelreq: string;
    cost: string;
    ['gamble cost']: string;
    code: string;
    namestr: string;
    ['magic lvl']: string;
    ['auto prefix']: string;
    alternategfx: string;
    OpenBetaGfx: string;
    normcode: string;
    ubercode: string;
    ultracode: string;
    spelloffset: string;
    component: string;
    invwidth: string;
    invheight: string;
    hasinv: string;
    gemsockets: string;
    gemapplytype: string;
    flippyfile: string;
    invfile: string;
    uniqueinvfile: string;
    setinvfile: string;
    rArm: string;
    lArm: string;
    Torso: string;
    Legs: string;
    rSPad: string;
    lSPad: string;
    useable: string;
    throwable: string;
    stackable: string;
    minstack: string;
    maxstack: string;
    type: string;
    type2: string;
    dropsound: string;
    dropsfxframe: string;
    usesound: string;
    unique: string;
    transparent: string;
    transtbl: string;
    quivered: string;
    lightradius: string;
    belt: string;
    quest: string;
    missiletype: string;
    durwarning: string;
    qntwarning: string;
    mindam: string;
    maxdam: string;
    StrBonus: string;
    DexBonus: string;
    gemoffset: string;
    bitfield1: string;
    CharsiMin: string;
    CharsiMax: string;
    CharsiMagicMin: string;
    CharsiMagicMax: string;
    CharsiMagicLvl: string;
    GheedMin: string;
    GheedMax: string;
    GheedMagicMin: string;
    GheedMagicMax: string;
    GheedMagicLvl: string;
    AkaraMin: string;
    AkaraMax: string;
    AkaraMagicMin: string;
    AkaraMagicMax: string;
    AkaraMagicLvl: string;
    FaraMin: string;
    FaraMax: string;
    FaraMagicMin: string;
    FaraMagicMax: string;
    FaraMagicLvl: string;
    LysanderMin: string;
    LysanderMax: string;
    LysanderMagicMin: string;
    LysanderMagicMax: string;
    LysanderMagicLvl: string;
    DrognanMin: string;
    DrognanMax: string;
    DrognanMagicMin: string;
    DrognanMagicMax: string;
    DrognanMagicLvl: string;
    HraltiMin: string;
    HraltiMax: string;
    HraltiMagicMin: string;
    HraltiMagicMax: string;
    HratliMagicLvl: string;
    AlkorMin: string;
    AlkorMax: string;
    AlkorMagicMin: string;
    AlkorMagicMax: string;
    AlkorMagicLvl: string;
    OrmusMin: string;
    OrmusMax: string;
    OrmusMagicMin: string;
    OrmusMagicMax: string;
    OrmusMagicLvl: string;
    ElzixMin: string;
    ElzixMax: string;
    ElzixMagicMin: string;
    ElzixMagicMax: string;
    ElzixMagicLvl: string;
    AshearaMin: string;
    AshearaMax: string;
    AshearaMagicMin: string;
    AshearaMagicMax: string;
    AshearaMagicLvl: string;
    CainMin: string;
    CainMax: string;
    CainMagicMin: string;
    CainMagicMax: string;
    CainMagicLvl: string;
    HalbuMin: string;
    HalbuMax: string;
    HalbuMagicMin: string;
    HalbuMagicMax: string;
    HalbuMagicLvl: string;
    JamellaMin: string;
    JamellaMax: string;
    JamellaMagicMin: string;
    JamellaMagicMax: string;
    JamellaMagicLvl: string;
    LarzukMin: string;
    LarzukMax: string;
    LarzukMagicMin: string;
    LarzukMagicMax: string;
    LarzukMagicLvl: string;
    MalahMin: string;
    MalahMax: string;
    MalahMagicMin: string;
    MalahMagicMax: string;
    MalahMagicLvl: string;
    DrehyaMin: string;
    DrehyaMax: string;
    DrehyaMagicMin: string;
    DrehyaMagicMax: string;
    DrehyaMagicLvl: string;
    ['Source Art']: string;
    ['Game Art']: string;
    Transform: string;
    InvTrans: string;
    SkipName: string;
    NightmareUpgrade: string;
    HellUpgrade: string;
    mindam_1: string;
    maxdam_2: string;
    nameable: string;
}