// src/types/UniqueItemData.ts
export interface UniqueItemData {
    index: string;
    version: string;
    enabled: string;
    ladder: string;
    rarity: string;
    nolimit: string;
    lvl: string;
    ['lvl req']: string;
    code: string;
    ['*type']: string;
    ['*uber']: string;
    carry1: string;
    ['cost mult']: string;
    ['cost add']: string;
    chrtransform: string;
    invtransform: string;
    flippyfile: string;
    invfile: string;
    dropsound: string;
    dropsfxframe: string;
    usesound: string;
    prop1: string;
    par1: string;
    min1: string;
    max1: string;
    prop2: string;
    par2: string;
    min2: string;
    max2: string;
    prop3: string;
    par3: string;
    min3: string;
    max3: string;
    prop4: string;
    par4: string;
    min4: string;
    max4: string;
    prop5: string;
    par5: string;
    min5: string;
    max5: string;
    prop6: string;
    par6: string;
    min6: string;
    max6: string;
    prop7: string;
    par7: string;
    min7: string;
    max7: string;
    prop8: string;
    par8: string;
    min8: string;
    max8: string;
    prop9: string;
    par9: string;
    min9: string;
    max9: string;
    prop10: string;
    par10: string;
    min10: string;
    max10: string;
    prop11: string;
    par11: string;
    min11: string;
    max11: string;
    prop12: string;
    par12: string;
    min12: string;
    max12: string;
    ['*eol']: string;
}